pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ansible>=11.6.0",
    "click>=8.2.1",
    "gitpython>=3.1.44",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
